Reorganized OpenXml issue tests
Moved issue tests in their own folder, separated gitee issues from github issues and ordered them, and moved models to own file.
